Vulkan renderer refactor with VMA and new pipeline system
Massive cleanup and rework of the Vulkan backend:
- Switched to VMA for memory allocation with proper tracking and stats
- Fixed leaks and improved GPU memory usage
- Introduced a new pipeline abstraction and render command system
- Replaced old RenderSystem with a modular, resource-driven architecture
- Added framebuffer, image, and buffer management

+++ Engine/Source/Zenith/Renderer/IndexBuffer.cpp
@@ -5,9 +5,11 @@
 
 #include "Zenith/Renderer/API/Vulkan/VulkanIndexBuffer.hpp"
 
+#include "Zenith/Renderer/RendererAPI.hpp"
+
 namespace Zenith {
 
-	Ref<IndexBuffer> IndexBuffer::Create(uint32_t size)
+	Ref<IndexBuffer> IndexBuffer::Create(uint64_t size)
 	{
 		switch (RendererAPI::Current())
 		{
@@ -18,7 +20,7 @@ namespace Zenith {
 		return nullptr;
 	}
 
-	Ref<IndexBuffer> IndexBuffer::Create(void* data, uint32_t size)
+	Ref<IndexBuffer> IndexBuffer::Create(void* data, uint64_t size)
 	{
 		switch (RendererAPI::Current())
 		{
