+++ Engine/Source/Zenith/Renderer/IndexBuffer.hpp
@@ -2,7 +2,7 @@
 
 #include "Zenith/Core/Ref.hpp"
 
-#include "RendererAPI.hpp"
+#include "RendererTypes.hpp"
 
 namespace Zenith {
 
@@ -11,16 +11,17 @@ namespace Zenith {
 	public:
 		virtual ~IndexBuffer() {}
 
-		virtual void SetData(void* buffer, uint32_t size, uint32_t offset = 0) = 0;
+		virtual void SetData(void* buffer, uint64_t size, uint64_t offset = 0) = 0;
 		virtual void Bind() const = 0;
 
 		virtual uint32_t GetCount() const = 0;
 
-		virtual uint32_t GetSize() const = 0;
+		virtual uint64_t GetSize() const = 0;
 		virtual RendererID GetRendererID() const = 0;
 
-		static Ref<IndexBuffer> Create(uint32_t size);
-		static Ref<IndexBuffer> Create(void* data, uint32_t size = 0);
+		static Ref<IndexBuffer> Create(uint64_t size);
+		static Ref<IndexBuffer> Create(void* data, uint64_t size = 0);
 	};
 
 }
+
